chore(refactor): rename renderers/collection-schema as ogc-api-feature/writer (refs #47)
